refactor(frontend): hoist CONVERSATION_DRAG_MIME → lib/conversations/drag
Bean pawrrtal-q8p8 Stage A. The drag-and-drop MIME was declared in
`features/projects/constants.ts` but consumed by
`features/nav-chats/components/ConversationSidebarItemView.tsx`
(the drag source). One feature reaching into another for a
shared protocol-level constant.

The constant moves to `frontend/lib/conversations/drag.ts` (new
directory) with its dedicated `drag.test.ts`. Three importers are
rewritten:

  - features/projects/components/ProjectRow.tsx
  - features/projects/components/ProjectRow.test.tsx
  - features/nav-chats/components/ConversationSidebarItemView.tsx

The MIME-specific case in `features/projects/constants.test.ts` is
removed (covered by the new `lib/conversations/drag.test.ts`); the
`PROJECTS_STORAGE_KEYS` test stays.

Stage B of the bean (extract `chat-activity-context` +
`use-conversation-mutations` to `lib/conversations/`) is deferred.

bun run typecheck + bun run arch:check: green.
features/projects + features/nav-chats + lib/conversations tests:
13/13 across 4 files.
