+++ frontend/features/projects/constants.ts
@@ -1,8 +1,10 @@
 /**
  * Constants for the projects feature.
  *
- * Keeps the localStorage key namespace + the mime type used by the
- * sidebar drag-and-drop in one place so renames don't drift across files.
+ * The conversation drag MIME used to live here too, but it's consumed by
+ * both `features/nav-chats/` (drag source) and `features/projects/` (drop
+ * target). It now lives at `@/lib/conversations/drag` to avoid a
+ * cross-feature edge.
  */
 
 /** localStorage keys owned by the projects feature. */
@@ -13,10 +15,3 @@ export const PROJECTS_STORAGE_KEYS = {
 
 /** Union of every recognized projects `localStorage` key. */
 export type ProjectsStorageKey = (typeof PROJECTS_STORAGE_KEYS)[keyof typeof PROJECTS_STORAGE_KEYS];
-
-/**
- * MIME type used in the HTML5 drag-and-drop dataTransfer when a chat row
- * is being dragged onto a project. Custom string so the drop target can
- * tell our payload apart from arbitrary text/uri-list drags.
- */
-export const CONVERSATION_DRAG_MIME = 'application/x-pawrrtal-conversation';
